refactor(frontend/controller): move search highlighter into seperate file
